+++ packages/twenty-server/src/modules/calendar/blocklist-manager/calendar-blocklist-manager.module.ts
@@ -4,10 +4,10 @@ import { BlocklistItemDeleteCalendarEventsJob } from 'src/modules/calendar/block
 import { BlocklistReimportCalendarEventsJob } from 'src/modules/calendar/blocklist-manager/jobs/blocklist-reimport-calendar-events.job';
 import { CalendarBlocklistListener } from 'src/modules/calendar/blocklist-manager/listeners/calendar-blocklist.listener';
 import { CalendarEventCleanerModule } from 'src/modules/calendar/calendar-event-cleaner/calendar-event-cleaner.module';
-import { CalendarEventImportManagerModule } from 'src/modules/calendar/calendar-event-import-manager/calendar-event-import-manager.module';
+import { CalendarCommonModule } from 'src/modules/calendar/common/calendar-common.module';
 
 @Module({
-  imports: [CalendarEventCleanerModule, CalendarEventImportManagerModule],
+  imports: [CalendarEventCleanerModule, CalendarCommonModule],
   providers: [
     CalendarBlocklistListener,
     BlocklistItemDeleteCalendarEventsJob,
