+++ packages/twenty-server/src/modules/calendar/common/calendar-common.module.ts
@@ -1,0 +1,18 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+
+import { FeatureFlagEntity } from 'src/engine/core-modules/feature-flag/feature-flag.entity';
+import { WorkspaceDataSourceModule } from 'src/engine/workspace-datasource/workspace-datasource.module';
+import { CalendarChannelSyncStatusService } from 'src/modules/calendar/common/services/calendar-channel-sync-status.service';
+import { ConnectedAccountModule } from 'src/modules/connected-account/connected-account.module';
+
+@Module({
+  imports: [
+    WorkspaceDataSourceModule,
+    TypeOrmModule.forFeature([FeatureFlagEntity], 'core'),
+    ConnectedAccountModule,
+  ],
+  providers: [CalendarChannelSyncStatusService],
+  exports: [CalendarChannelSyncStatusService],
+})
+export class CalendarCommonModule {}
